requirements.txt
python-binance
pandas
ta
numpy
requests
python-dotenv